wip: After testing in the f1tenth_gym
The car is doing a lap in ~50sec, which is ~15sec better than before [!].
I used the .pkl file with GradientDriver and needed to add three attributes to it (they are in the GradientDriver.plan method), they need to be added to the GradientDriver.__init__ signature).

Overall, this is a checkpoint for something that seems to work!